refactor(mem-pool): move custodian code to block-producer and utils
