+++ crates/mem-pool/src/constants.rs
@@ -2,5 +2,3 @@
 pub const MIN_CKB_DEPOSIT_CAPACITY: u64 = 298_00000000;
 /// MIN Simple UDT deposit capacity, calculated from custodian cell size + simple UDT script
 pub const MIN_SUDT_DEPOSIT_CAPACITY: u64 = 379_00000000;
-/// MAX custodian cells
-pub const MAX_CUSTODIANS: usize = 50;
