+++ crates/tests/src/tests/deprecated/helper.rs
@@ -5,10 +5,6 @@ use gw_block_producer::custodian::AvailableCustodians;
 use gw_common::H256;
 use gw_config::ContractsCellDep;
 use gw_generator::error::WithdrawalError;
-use gw_mem_pool::custodian::{
-    build_finalized_custodian_lock, calc_ckb_custodian_min_capacity, generate_finalized_custodian,
-    sum_withdrawals,
-};
 use gw_types::{
     bytes::Bytes,
     core::ScriptHashType,
@@ -18,6 +14,13 @@ use gw_types::{
         WithdrawalRequestExtra,
     },
     prelude::*,
+};
+use gw_utils::{
+    custodian::{
+        build_finalized_custodian_lock, calc_ckb_custodian_min_capacity,
+        generate_finalized_custodian,
+    },
+    withdrawal::sum_withdrawals,
 };
 
 use std::collections::HashMap;
